+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-Current
--------
+0.1.2
+-----
 
 - Added ``Api.marshal_with()`` and ``Api.marshal_list_with()`` decorators
 - Added ``Api.marshal()`` shortcut